Add zAsar-mode.el for editing zAsar 65816 Assembly code
This commit adds the zAsar-mode.el file, which provides a major mode for editing zAsar 65816 Assembly code in Emacs. The mode includes syntax highlighting, code navigation, and other features. It also defines a keymap and a menu for easy access to various commands related to zAsar. Additionally, it includes parsing functions to parse labels from disassembly files and the current project, as well as functions to list references and jump to the definition of labels.